components/dbloader/pyproject.toml
[project]
name = "dbloader"
dependencies = [
    "typer>=0.12.5",
    "srt>=3.5.3",
    "python-dotenv>=1.0.1",
    "pymongo>=4.10.1",
    "requests>=2.32.3",
    "pysolr>=3.10.0",
    "boto3>=1.35.49",
    "jsonschema>=4.23.0",
    "pyyaml>=6.0.2",
    "pytz>=2024.2",
    "fastapi[standard]>=0.115.4",
]
